rename snapshot_base -> snapshot_reader
