+++ silkworm/db/snapshots/header_snapshot.hpp
@@ -19,7 +19,7 @@
 #include <silkworm/core/common/bytes.hpp>
 #include <silkworm/core/types/block.hpp>
 
-#include "snapshot_base.hpp"
+#include "snapshot_reader.hpp"
 #include "snapshot_word_serializer.hpp"
 
 namespace silkworm::snapshots {
